# apps/api/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Inventario API",
    "description": "REST API for Inventario — a multi-tenant, white-label asset management platform for sports federations, municipalities, schools, and clubs. Open-source under EUPL-1.2.",
    "version": "0.1.0",
    "contact": {
      "name": "Inventario · LTK Solutions",
      "url": "https://inventario.sportup.sk",
      "email": "inventario@sportup.sk"
    },
    "license": {
      "name": "EUPL-1.2",
      "url": "https://joinup.ec.europa.eu/collection/eupl/eupl-text-eupl-12"
    }
  },
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "description": "Microsoft Entra ID access token (v2.0). Multi-tenant: the JWT `tid` claim resolves the Organisation; the `oid` claim resolves the User. See apps/api/README.md for the device code flow."
      },
      "deploymentSecret": {
        "type": "http",
        "scheme": "bearer",
        "description": "Zdieľané tajomstvo pre servisné endpointy pod /v1/system — nie používateľský token. Migrácie a indexy používajú MIGRATIONS_SECRET (volá GitHub Actions po produkčnom deployi), retencia CRON_SECRET (volá Vercel Cron). Bez nastavenej premennej endpoint vracia 503."
      }
    },
...
  },
  "paths": {
...
    "/v1/system/migrations/run": {
      "post": {
        "operationId": "postV1SystemMigrationsRun",
        "summary": "Spustenie čakajúcich DB migrácií po deployi",
        "tags": [
          "System"
        ],
        "description": "Volá GitHub Actions po úspešnom produkčnom deployi. Cold start migrácie nespúšťa, robí len pasívnu kontrolu. Opakované spustenie je bezpečné.",
        "security": [
          {
            "deploymentSecret": []
          }
        ],
        "responses": {
          "200": {
            "description": "Default Response"
          }
        }
      }
    },
...
    "/v1/system/retention/run": {
      "post": {
        "operationId": "postV1SystemRetentionRun",
        "summary": "Mesačné uplatnenie retenčnej politiky",
        "tags": [
          "System"
        ],
        "description": "Volá Vercel Cron (mesačne). Maže a pseudonymizuje dáta podľa docs/compliance/data-retention-schedule.md.",
        "security": [
          {
            "deploymentSecret": []
          }
        ],
        "responses": {
          "200": {
            "description": "Default Response"
          }
        }
      }
    },
...
  },
  "servers": [
    {
      "url": "http://localhost:3000",
      "description": "Local dev"
    },
    {
      "url": "https://api.inventario.sportup.sk",
      "description": "Production (planned Q3 2026)"
    }
  ],
...
}